[DELETE] - mock_pyglet.py [EDIT] - mock_turtle.py - Merged in contents of pyglet.py. [EDIT] - code_tracer.py - Importing MockPyglet from mock_turtle module.
